requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cvzone==1.6.1
cycler==0.12.1
Django==5.1
flatbuffers==24.3.25
fonttools==4.53.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
google-search-results==2.4.2
googleapis-common-protos==1.63.2
grpcio==1.65.5
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.7
jax==0.4.31
jaxlib==0.4.31
kiwisolver==1.4.5
matplotlib==3.9.2
mediapipe==0.10.14
ml-dtypes==0.4.0
numpy==2.1.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
pyvista==0.44.1
requests==2.32.3
rsa==4.9
scipy==1.14.1
scooby==0.10.0
six==1.16.0
sounddevice==0.5.0
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vtk==9.3.1
